requirements.txt
streamlit
google-genai
pandas
pyreadstat
xlsxwriter